Add --run-list and --json-attributes options to generate node config
This commit adds `--run-list` and `--json-attributes` options to `knife
solo bootstrap`, `knife prepare` and `knife cook`. The options and their
format is the same than in for example `knife bootstrap`. But they are
only used when generating a new node configuration file.

This allows `knife bootstrap` to be replaced with `knife solo
bootstrap`,  making it possible to inject it to e.g. knife-ec2 plugin.

Fixes #108.